test: Lotto, LottoMachine 테스트
