+++ src/main/kotlin/lotto/model/Lotto.kt
@@ -6,7 +6,9 @@ class Lotto(private val numbers: List<Int>) {
     init {
         require(numbers.size == LOTTO_SIZE) { "[ERROR] 로또 번호는 ${LOTTO_SIZE}개여야 합니다" }
         require(numbers.distinct().size == LOTTO_SIZE) { "[ERROR] 로또 번호는 중복되지 않아야 합니다" }
-        require(numbers.all { it in NUMBER_RANGE }) { "[ERROR] 로또 번호는 ${MIN_NUMBER}부터 ${MAX_NUMBER}사이의 숫자여야 합니다" }
+        require(numbers.all { it in LottoNumber.NUMBER_RANGE }) {
+            "[ERROR] 로또 번호는 ${LottoNumber.MIN_NUMBER}부터 ${LottoNumber.MAX_NUMBER}사이의 숫자여야 합니다"
+        }
     }
 
     fun containNumber(lottoNumber: LottoNumber): Boolean {
@@ -21,8 +23,5 @@ class Lotto(private val numbers: List<Int>) {
 
     companion object {
         const val LOTTO_SIZE = 6
-        const val MIN_NUMBER = 1
-        const val MAX_NUMBER = 45
-        val NUMBER_RANGE = MIN_NUMBER..MAX_NUMBER
     }
 }
