+++ src/main/kotlin/lotto/model/LottoMachine.kt
@@ -4,12 +4,12 @@ import camp.nextstep.edu.missionutils.Randoms
 
 object LottoMachine {
 
-    fun issueLotto(purchaseAmount: PurchaseAmount): List<Lotto> {
+    fun issueLottos(purchaseAmount: PurchaseAmount): List<Lotto> {
         val lottoCount = purchaseAmount.getLottoCount()
         val lottos = mutableListOf<Lotto>()
         repeat(lottoCount) {
             lottos.add(
-                Lotto(Randoms.pickUniqueNumbersInRange(Lotto.MIN_NUMBER, Lotto.MAX_NUMBER, Lotto.LOTTO_SIZE))
+                Lotto(Randoms.pickUniqueNumbersInRange(LottoNumber.MIN_NUMBER, LottoNumber.MAX_NUMBER, Lotto.LOTTO_SIZE))
             )
         }
         return lottos
